# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Night City RP Portal API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /stores/{id}:
    parameters:
      - { in: path, name: id, required: true, schema: { type: integer } }
    get:
      operationId: getStore
      tags: [stores]
      responses:
        "200":
          description: Store with stock
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Store" }
    patch:
      operationId: updateStore
      tags: [stores]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/StoreUpdate" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Store" }
  /stores/{id}/employees:
    parameters:
      - { in: path, name: id, required: true, schema: { type: integer } }
    post:
      operationId: addStoreEmployee
      tags: [stores]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/EmployeeInput" }
      responses:
        "201":
          description: Added
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Employee" }
  /stores/{id}/employees/{employeeId}:
    parameters:
      - { in: path, name: id, required: true, schema: { type: integer } }
      - { in: path, name: employeeId, required: true, schema: { type: integer } }
    delete:
      operationId: removeStoreEmployee
      tags: [stores]
      responses:
        "204": { description: Removed }
  /stores/{id}/stock:
    parameters:
      - { in: path, name: id, required: true, schema: { type: integer } }
    post:
      operationId: addStoreStock
      tags: [stores]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/StockInput" }
      responses:
        "201":
          description: Added
          content:
            application/json:
              schema: { $ref: "#/components/schemas/StockItem" }
  /stores/{id}/stock/{stockId}:
    parameters:
      - { in: path, name: id, required: true, schema: { type: integer } }
      - { in: path, name: stockId, required: true, schema: { type: integer } }
    patch:
      operationId: updateStoreStock
      tags: [stores]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/StockUpdate" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/StockItem" }
    delete:
      operationId: removeStoreStock
      tags: [stores]
      responses:
        "204": { description: Removed }

  # ---------------- RIPPERDOC MANAGEMENT ----------------
...
  /ripperdocs/{id}:
    parameters:
      - { in: path, name: id, required: true, schema: { type: integer } }
    get:
      operationId: getRipperdoc
      tags: [ripperdocs]
      responses:
        "200":
          description: Clinic with stock
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Ripperdoc" }
    patch:
      operationId: updateRipperdoc
      tags: [ripperdocs]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RipperdocUpdate" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Ripperdoc" }
  /ripperdocs/{id}/employees:
    parameters:
      - { in: path, name: id, required: true, schema: { type: integer } }
    post:
      operationId: addRipperdocEmployee
      tags: [ripperdocs]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/EmployeeInput" }
      responses:
        "201":
          description: Added
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Employee" }
  /ripperdocs/{id}/employees/{employeeId}:
    parameters:
      - { in: path, name: id, required: true, schema: { type: integer } }
      - { in: path, name: employeeId, required: true, schema: { type: integer } }
    delete:
      operationId: removeRipperdocEmployee
      tags: [ripperdocs]
      responses:
        "204": { description: Removed }
  /ripperdocs/{id}/stock:
    parameters:
      - { in: path, name: id, required: true, schema: { type: integer } }
    post:
      operationId: addRipperdocStock
      tags: [ripperdocs]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/StockInput" }
      responses:
        "201":
          description: Added
          content:
            application/json:
              schema: { $ref: "#/components/schemas/StockItem" }
  /ripperdocs/{id}/stock/{stockId}:
    parameters:
      - { in: path, name: id, required: true, schema: { type: integer } }
      - { in: path, name: stockId, required: true, schema: { type: integer } }
    delete:
      operationId: removeRipperdocStock
      tags: [ripperdocs]
      responses:
        "204": { description: Removed }

  # ---------------- FIXER HUB ----------------
...
  /sheets:
    get:
      operationId: listMySheets
      tags: [sheets]
      summary: My submitted character sheets
      responses:
        "200":
          description: Sheets
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/CharacterSheet" }
    post:
      operationId: submitSheet
      tags: [sheets]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CharacterSheetInput" }
      responses:
        "201":
          description: Submitted; CS Approvers notified on Discord
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CharacterSheet" }
...
  /sheets/{id}:
    parameters:
      - { in: path, name: id, required: true, schema: { type: integer } }
    get:
      operationId: getSheet
      tags: [sheets]
      responses:
        "200":
          description: Sheet
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CharacterSheet" }
...
components:
  schemas:
...
    Store:
      type: object
      required: [id, name, kind, ownerId, employees, stock]
      properties:
        id: { type: integer }
        name: { type: string }
        ownerId: { type: string }
        ownerCharacterId: { type: ["integer", "null"] }
        kind: { type: string, enum: [guns, gear, clothing, mixed, other] }
        location: { type: ["string", "null"] }
        description: { type: ["string", "null"] }
        bannerUrl: { type: ["string", "null"] }
        employees:
          type: array
          items: { $ref: "#/components/schemas/Employee" }
        stock:
          type: array
          items: { $ref: "#/components/schemas/StockItem" }
    StoreUpdate:
      type: object
      properties:
        name: { type: string }
        kind: { type: string, enum: [guns, gear, clothing, mixed, other] }
        location: { type: string }
        description: { type: string }
        bannerUrl: { type: string }
    Ripperdoc:
      type: object
      required: [id, name, ownerId, employees, stock]
      properties:
        id: { type: integer }
        name: { type: string }
        ownerId: { type: string }
        ownerCharacterId: { type: ["integer", "null"] }
        location: { type: ["string", "null"] }
        description: { type: ["string", "null"] }
        bannerUrl: { type: ["string", "null"] }
        employees:
          type: array
          items: { $ref: "#/components/schemas/Employee" }
        stock:
          type: array
          items: { $ref: "#/components/schemas/StockItem" }
    RipperdocUpdate:
      type: object
      properties:
        name: { type: string }
        location: { type: string }
        description: { type: string }
        bannerUrl: { type: string }
    Employee:
      type: object
      required: [id, characterId, name, role]
      properties:
        id: { type: integer }
        characterId: { type: integer }
        name: { type: string }
        role: { type: string }
    EmployeeInput:
      type: object
      required: [characterId, role]
      properties:
        characterId: { type: integer }
        role: { type: string, minLength: 1 }
    StockItem:
      type: object
      required: [id, name, price, quantity]
      properties:
        id: { type: integer }
        name: { type: string }
        category: { type: ["string", "null"] }
        price: { type: integer }
        quantity: { type: integer }
        notes: { type: ["string", "null"] }
    StockInput:
      type: object
      required: [name, price, quantity]
      properties:
        name: { type: string, minLength: 1 }
        category: { type: string }
        price: { type: integer, minimum: 0 }
        quantity: { type: integer, minimum: 0 }
        notes: { type: string }
    StockUpdate:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        category: { type: string }
        price: { type: integer, minimum: 0 }
        quantity: { type: integer, minimum: 0 }
        notes: { type: string }
...
    CharacterSheet:
      type: object
      required: [id, ownerId, name, status, createdAt, data]
      properties:
        id: { type: integer }
        ownerId: { type: string }
        ownerName: { type: string }
        characterId: { type: ["integer", "null"] }
        name: { type: string }
        status: { type: string, enum: [pending, approved, rejected, changes_requested] }
        decisionBy: { type: ["string", "null"] }
        decisionNote: { type: ["string", "null"] }
        decidedAt: { type: ["string", "null"], format: date-time }
        createdAt: { type: string, format: date-time }
        data: { $ref: "#/components/schemas/CharacterSheetData" }
    CharacterSheetData:
      type: object
      required: [fullName, archetype, age, gender, appearance, background, attributes, skills, cyberware, gear, startingEddies]
      properties:
        fullName: { type: string }
        nickname: { type: ["string", "null"] }
        archetype: { type: string }
        age: { type: integer }
        gender: { type: string }
        appearance: { type: string }
        background: { type: string }
        attributes:
          type: object
          additionalProperties: { type: integer }
        skills:
          type: object
          additionalProperties: { type: integer }
        cyberware:
          type: array
          maxItems: 11
          items: { $ref: "#/components/schemas/CharacterSheetCyberwareEntry" }
        cyberwarePointsSpent: { type: integer, maximum: 6 }
        gear:
          type: array
          items: { type: string }
        startingEddies: { type: integer }
        notes: { type: ["string", "null"] }
    CharacterSheetCyberwareEntry:
      type: object
      required: [slot, name, points]
      properties:
        slot: { type: string }
        name: { type: string }
        points: { type: integer }
        humanityLoss: { type: integer }
        notes: { type: ["string", "null"] }
    CharacterSheetInput:
      type: object
      required: [name, data]
      properties:
        name: { type: string, minLength: 1, maxLength: 64 }
        characterId: { type: ["integer", "null"] }
        data: { $ref: "#/components/schemas/CharacterSheetData" }
...